chore(ui): clippy-clean kael_ui and add it to 3-OS CI verification
- fix all clippy lints (manual clamp, checked div, identical branches,
  Rc instead of non-Send Arc for validation closures, ptr_arg, sort_by_key,
  duplicate AccordionItem builder, no-op rating listener)
- allow type_complexity/too_many_arguments crate-wide (callback-heavy UI API)
- CI: verify-kael.sh/.ps1 now lint, test, and type-check kael_ui, all
  examples, and the three template apps on macOS, Windows, and Linux

+++ crates/kael_ui/src/components/progress.rs
@@ -350,7 +350,7 @@ impl RenderOnce for CircularProgress {
                                         .border(stroke_w)
                                         .border_color(stroke_color)
                                         .rounded(px(9999.0))
-                                        .opacity((value * 0.7 + 0.3) as f32),
+                                        .opacity(value * 0.7 + 0.3),
                                 )
                                 .into_any_element(),
                             (None, SpinnerType::Dot) => {
